Chagned the names of file and thought of a good idea to improve the performance
