+++ source/rust/src/lib.rs
@@ -1,3 +1,3 @@
 #[allow(non_snake_case)]
 
-mod rect;
+mod Recognition;
